--- openstack/blockstorage/v3/extensions/quotasets/testing/doc.go
@@ -1,2 +1,0 @@
-// quotasets unit tests
-package testing
